Feat: moved set implementation to DS package
