+++ ds/randommap/random_map_test.go
@@ -6,8 +6,8 @@ import (
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 
+	"github.com/iotaledger/hive.go/ds"
 	"github.com/iotaledger/hive.go/ds/randommap"
-	"github.com/iotaledger/hive.go/ds/set"
 )
 
 func TestRandomMap_Basics(t *testing.T) {
@@ -92,5 +92,5 @@ func TestRandomMap_EmptyMap(t *testing.T) {
 }
 
 func containsUniqueElements[V comparable](list []V) bool {
-	return set.New[V](list...).Size() == len(list)
+	return ds.NewSet[V](list...).Size() == len(list)
 }
